hotfix revenue calculation

+++ common/models/Restaurant.php
@@ -2127,7 +2127,7 @@ class Restaurant extends \yii\db\ActiveRecord
 
         return [
             'revenue_generated_chart_data' => array_values($revenue_generated_chart_data),
-            'number_of_all_revenue_generated' => (int) $number_of_all_revenue_generated
+            'number_of_all_revenue_generated' => (float) $number_of_all_revenue_generated
         ];
     }
 
@@ -2283,7 +2283,7 @@ class Restaurant extends \yii\db\ActiveRecord
 
         return [
             'revenue_generated_chart_data' => array_values($revenue_generated_chart_data),
-            'number_of_all_revenue_generated' => (int) $number_of_all_revenue_generated
+            'number_of_all_revenue_generated' => (float) $number_of_all_revenue_generated
         ];
     }
 
@@ -2447,7 +2447,7 @@ class Restaurant extends \yii\db\ActiveRecord
 
         return [
             'revenue_generated_chart_data' => array_values($revenue_generated_chart_data),
-            'number_of_all_revenue_generated' => (int) $number_of_all_revenue_generated
+            'number_of_all_revenue_generated' => (float) $number_of_all_revenue_generated
         ];
     }
